Make RandomBeatable use RandomAssumed Algorithm.
Update the RandomBeatable filler to be based on the less biased RandomAssumed algorithm.
It is now basically equivalent to Entrance Randomizer's `--only-beatable` option.

Basically whenever we are about to place an item, we will check if the game would be beatable without that item (taking into account where existing items were placed). If the game would be beatable without it, we allow it to be placed anywhere, subject only to Fill Rule constraints. (i.e. without checking if the location would be accessible).